Outsource map click into ol/Interaction subclass
